+++ tests/test_csc.py
@@ -41,6 +41,7 @@ class TestM2CSC(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
         )
 
     def setUp(self):
+        super().setUp()
         self.csc_mtmount = None
 
     async def _simulate_csc_mount(self, elevation_angle):
@@ -55,8 +56,8 @@ class TestM2CSC(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
         self.csc_mtmount = salobj.Controller("MTMount")
         await self.csc_mtmount.start_task
 
-        self.csc_mtmount.tel_elevation.set_put(actualPosition=elevation_angle)
-        self.csc_mtmount.evt_elevationInPosition.set_put(inPosition=True)
+        await self.csc_mtmount.tel_elevation.set_write(actualPosition=elevation_angle)
+        await self.csc_mtmount.evt_elevationInPosition.set_write(inPosition=True)
 
         # Wait for some time to publish the telemetry of MTMount
         await asyncio.sleep(1)
@@ -379,6 +380,29 @@ class TestM2CSC(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
                     await getattr(self.remote, f"tel_{tel}").next(
                         flush=True, timeout=STD_TIMEOUT
                     )
+
+    async def test_connection_monitor_loop(self):
+        async with self.make_csc(
+            initial_state=salobj.State.STANDBY, config_dir=None, simulation_mode=1
+        ):
+
+            # Enter the Disabled state to construct the connection
+            await salobj.set_summary_state(self.remote, salobj.State.DISABLED)
+            self.assertTrue(self.csc.model.are_clients_connected())
+
+            time_wait_connection_monitor_check = 2
+            await asyncio.sleep(time_wait_connection_monitor_check)
+
+            await self.csc._mock_server.close()
+
+            # Sleep some time to make sure the CSC detects the connection is
+            # closed
+            await asyncio.sleep(time_wait_connection_monitor_check)
+
+            self.assertFalse(self.csc.model.are_clients_connected())
+            self.assertFalse(self.csc._mock_server.are_servers_connected())
+
+            self.assertEqual(self.csc.summary_state, salobj.State.FAULT)
 
     async def test_applyForces_wrong_controller_state(self):
         async with self.make_csc(