+++ doc/version_history.rst
@@ -2,6 +2,25 @@
 Version History
 ===============
 
+v0.5.2
+------
+
+* Actively monitor the connection status.
+If the server closes the connection, M2 CSC will detect this and transition
+to the **Fault** state.
+If there is no new telemetry for some time, there will be the warning message.
+
+v0.5.1
+------
+
+* Update the mechanism to overwrite the connection information.
+
+v0.5.0
+------
+
+* Update to salobj 7.
+* Rename **README.rst** to **README.md** and update the related syntax.
+
 v0.4.1
 ------
 * In `CSC`: